# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Median Code API
  description: |
    Production-ready FastAPI code generation and API entity management with Clerk JWT authentication.

    ## Authentication

    All endpoints require Bearer JWT authentication via Clerk. Include your Clerk JWT token in the Authorization header:

    ```
    Authorization: Bearer <your-clerk-jwt-token>
    ```

    ## API Overview

    This API provides two main categories of endpoints:

    ### 1. Entity Management (CRUD Operations)

    Manage your API design entities with full CRUD operations:

    - **Namespaces** (`/namespaces`) - Create and manage namespaces for organizing entities
    - **APIs** (`/apis`) - Create and manage API definitions that group endpoints and tags
    - **Types** (`/types`) - Retrieve primitive and abstract type definitions (read-only)
    - **Validators** (`/validators`) - Retrieve validator definitions (read-only)
    - **Fields** (`/fields`) - Create, read, update, and delete field definitions
    - **Objects** (`/objects`) - Create, read, update, and delete object definitions
    - **Tags** (`/tags`) - Create, read, update, and delete endpoint tags
    - **Endpoints** (`/endpoints`) - Create, read, update, and delete API endpoint definitions

    All list endpoints support filtering by `namespaceId` query parameter.

    ### 2. Code Generation

...

    #### Generated Application Structure

    The generated zip file contains:
...
    - `requirements.txt` - Python dependencies
...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /apis:
    get:
      tags:
        - APIs
      summary: List all APIs
      description: Retrieve all API definitions accessible to the authenticated user. APIs can be filtered by namespace.
      operationId: listApis
      parameters:
        - name: namespaceId
          in: query
          required: false
          schema:
            type: string
          description: Filter APIs by namespace ID
      responses:
        '200':
          description: Successfully retrieved APIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Api'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - APIs
      summary: Create a new API
      description: Create a new API definition. Server automatically sets createdAt and updatedAt timestamps.
      operationId: createApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - namespaceId
                - title
                - version
              properties:
                namespaceId:
                  type: string
                  example: "namespace-user"
                title:
                  type: string
                  example: "User Management API"
                version:
                  type: string
                  example: "1.0.0"
                description:
                  type: string
                  example: "API for managing user accounts"
                baseUrl:
                  type: string
                  example: "/api/v1"
                serverUrl:
                  type: string
                  example: "https://api.example.com"
      responses:
        '201':
          description: API created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /apis/{id}:
    get:
      tags:
        - APIs
      summary: Get API by ID
      description: Retrieve a specific API definition by its ID.
      operationId: getApi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API ID
      responses:
        '200':
          description: Successfully retrieved API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api'
        '404':
          description: API not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - APIs
      summary: Update API
      description: Update an existing API definition. Server automatically updates the updatedAt timestamp.
      operationId: updateApi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated API Title"
                version:
                  type: string
                  example: "2.0.0"
                description:
                  type: string
                  example: "Updated API description"
                baseUrl:
                  type: string
                  example: "/api/v2"
                serverUrl:
                  type: string
                  example: "https://api.example.com"
      responses:
        '200':
          description: API updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: API not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - APIs
      summary: Delete API
      description: Delete an API definition. This operation cascades and also deletes all endpoints and tags belonging to this API.
      operationId: deleteApi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API ID
      responses:
        '204':
          description: API deleted successfully (including all endpoints and tags)
        '404':
          description: API not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /apis/{id}/generate:
    post:
      tags:
        - APIs
      summary: Generate FastAPI application
      description: |
        Generates a complete, production-ready FastAPI application for the specified API.

        The generated application includes:
        - Fully configured FastAPI app with all endpoints
        - Pydantic models for validation
        - Type hints throughout
        - OpenAPI documentation
        - Requirements file with dependencies
        - README with setup instructions

        The server automatically retrieves all related entities (tags, endpoints, objects, fields, validators) for the specified API ID.

        **Returns:** A binary zip file containing the complete application structure.

        **Authentication:** Requires valid Clerk JWT token.
      operationId: generateApiCode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API ID to generate code for
      responses:
        '200':
          description: Successfully generated FastAPI application
          content:
            application/zip:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="fastapi-app.zip"
        '404':
          description: API not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "API with ID 'api-1' not found"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token from authentication flow
  schemas:
...
    Api:
      type: object
      required:
        - id
        - namespaceId
        - title
        - version
        - description
        - baseUrl
        - serverUrl
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Client-generated unique identifier
          example: "api-1"
        namespaceId:
          type: string
          description: Namespace this API belongs to
          example: "namespace-global"
        title:
          type: string
          description: API title for OpenAPI spec
          example: "User Management API"
        version:
          type: string
          description: API version (semantic versioning)
          example: "1.0.0"
        description:
          type: string
          description: API description
          example: "Endpoints for user management"
        baseUrl:
          type: string
          description: Base path for all endpoints
          example: "/api/v1"
        serverUrl:
          type: string
          description: Full server URL
          example: "https://api.example.com"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the API was created
          example: "2026-01-25T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the API was last updated
          example: "2026-01-25T10:30:00Z"

    # ========================================================================
    # Type Schema
    # ========================================================================
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Invalid API specification: endpoint path must start with /"
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing Clerk JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authenticated"